test: move ALObject test to new file
